requirements.txt
django-bootstrap4==1.1.1
djangorestframework==3.11.0
Pillow==6.2.1
psycopg2==2.8.4
django-extensions==2.2.9
django-bootstrap4==1.1.1
docutils==0.16
django-nested-admin==3.3.1